refactor(engine): organize database registration
